feat: redesigned course route structure
